+++ tests/unittests/OperatorTest.cpp
@@ -128,7 +128,7 @@ TEST(Operator, broadcast) {
   auto broadcastedBHandle = broadcasted->getPayload().getHandle();
   // Verify broadcasted B has same shape.
   EXPECT_EQ(numDims_A, broadcastedBHandle.dims().size());
-  for (int i = 0; i < broadcastedBHandle.dims().size(); i++) {
+  for (size_t i = 0; i < broadcastedBHandle.dims().size(); i++) {
     EXPECT_EQ(dims_A[i], broadcastedBHandle.dims()[i]);
   }
